refactor(backend): split auth api.go into domain-specific handler files
Extract handlers from api/auth/api.go (1,962 lines) into focused modules:
- session_handlers.go: login, register, refresh, logout
- password_handlers.go: change password, password reset
- account_handlers.go: account CRUD, activate/deactivate
- role_handlers.go: role CRUD and assignments
- permission_handlers.go: permission CRUD and assignments
- token_handlers.go: token management
- parent_account_handlers.go: parent account CRUD
- test_handlers.go: exported handler getters for testing

Reduces api.go from 1,962 to 220 lines. No functionality changes.

Refs: #494
