+++ backend/api/auth/api.go
@@ -1,25 +1,17 @@
 package auth
 
 import (
-	"database/sql"
-	"errors"
-	"log"
 	"net"
 	"net/http"
-	"strconv"
+	"regexp"
 	"strings"
-	"time"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/jwtauth/v5"
 	"github.com/go-chi/render"
-	validation "github.com/go-ozzo/ozzo-validation"
-	"github.com/go-ozzo/ozzo-validation/is"
 
-	"github.com/moto-nrw/project-phoenix/api/common"
 	"github.com/moto-nrw/project-phoenix/auth/authorize"
 	"github.com/moto-nrw/project-phoenix/auth/jwt"
-	authModel "github.com/moto-nrw/project-phoenix/models/auth"
 	authService "github.com/moto-nrw/project-phoenix/services/auth"
 )
 
@@ -183,1545 +175,6 @@ func (rs *Resource) Router() chi.Router {
 	return r
 }
 
-// LoginRequest represents the login request payload
-type LoginRequest struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-// Bind validates the login request
-func (req *LoginRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-		validation.Field(&req.Password, validation.Required),
-	)
-}
-
-// TokenResponse represents the token response payload
-type TokenResponse struct {
-	AccessToken  string `json:"access_token"`
-	RefreshToken string `json:"refresh_token"`
-}
-
-// login handles user login
-func (rs *Resource) login(w http.ResponseWriter, r *http.Request) {
-	req := &LoginRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	// Get IP address and user agent for audit logging
-	ipAddress := getClientIP(r)
-	userAgent := r.Header.Get(headerUserAgent)
-
-	accessToken, refreshToken, err := rs.AuthService.LoginWithAudit(r.Context(), req.Email, req.Password, ipAddress, userAgent)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidCredentials):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrInvalidCredentials))
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrInvalidCredentials)) // Mask the specific error
-			case errors.Is(authErr.Err, authService.ErrAccountInactive):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrAccountInactive))
-			default:
-				common.RenderError(w, r, ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	// Special case for login endpoint - frontend expects direct token response
-	render.JSON(w, r, TokenResponse{
-		AccessToken:  accessToken,
-		RefreshToken: refreshToken,
-	})
-}
-
-// RegisterRequest represents the register request payload
-type RegisterRequest struct {
-	Email           string `json:"email"`
-	Username        string `json:"username"`
-	Password        string `json:"password"`
-	ConfirmPassword string `json:"confirm_password"`
-	RoleID          *int64 `json:"role_id,omitempty"` // Optional role assignment
-}
-
-// Bind validates the register request
-func (req *RegisterRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-	req.Username = strings.TrimSpace(req.Username)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-		validation.Field(&req.Username, validation.Required, validation.Length(3, 30)),
-		validation.Field(&req.Password, validation.Required, validation.Length(8, 0)),
-		validation.Field(&req.ConfirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if req.Password != req.ConfirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// AccountResponse represents the account response payload
-type AccountResponse struct {
-	ID          int64    `json:"id"`
-	Email       string   `json:"email"`
-	Username    string   `json:"username,omitempty"`
-	Active      bool     `json:"active"`
-	Roles       []string `json:"roles,omitempty"`
-	Permissions []string `json:"permissions,omitempty"`
-}
-
-// register handles user registration
-func (rs *Resource) register(w http.ResponseWriter, r *http.Request) {
-	req := &RegisterRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	// Authorize role assignment (if role_id specified)
-	roleID, shouldReturn := rs.authorizeRoleAssignment(w, r, req.RoleID)
-	if shouldReturn {
-		return
-	}
-
-	account, err := rs.AuthService.Register(r.Context(), req.Email, req.Username, req.Password, roleID)
-	if err != nil {
-		rs.handleRegistrationError(w, r, err)
-		return
-	}
-
-	resp := buildAccountResponse(account)
-	common.Respond(w, r, http.StatusCreated, resp, "Account registered successfully")
-}
-
-// authorizeRoleAssignment checks if the caller is authorized to assign a role during registration.
-// Returns the authorized role ID and a boolean indicating if the handler should return early.
-func (rs *Resource) authorizeRoleAssignment(w http.ResponseWriter, r *http.Request, requestedRoleID *int64) (*int64, bool) {
-	if requestedRoleID == nil {
-		return nil, false
-	}
-
-	authHeader := r.Header.Get("Authorization")
-	if !isValidAuthHeader(authHeader) {
-		log.Printf("Security: Unauthenticated register attempt with role_id, ignoring role_id")
-		return nil, false
-	}
-
-	token := authHeader[7:]
-	callerAccount, err := rs.AuthService.ValidateToken(r.Context(), token)
-	if err != nil {
-		log.Printf("Security: Invalid token in register with role_id, ignoring role_id")
-		return nil, false
-	}
-
-	if !hasAdminRole(callerAccount.Roles) {
-		log.Printf("Security: Non-admin (account %d) attempted to set role_id, denying", callerAccount.ID)
-		common.RenderError(w, r, ErrorUnauthorized(errors.New("only administrators can assign roles")))
-		return nil, true
-	}
-
-	return requestedRoleID, false
-}
-
-// isValidAuthHeader checks if the Authorization header contains a valid Bearer token format
-func isValidAuthHeader(authHeader string) bool {
-	return authHeader != "" && len(authHeader) >= 8 && authHeader[:7] == "Bearer "
-}
-
-// hasAdminRole checks if any of the roles has the "admin" name
-func hasAdminRole(roles []*authModel.Role) bool {
-	for _, role := range roles {
-		if role.Name == "admin" {
-			return true
-		}
-	}
-	return false
-}
-
-// handleRegistrationError handles authentication errors during registration
-func (rs *Resource) handleRegistrationError(w http.ResponseWriter, r *http.Request, err error) {
-	var authErr *authService.AuthError
-	if !errors.As(err, &authErr) {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	switch {
-	case errors.Is(authErr.Err, authService.ErrEmailAlreadyExists):
-		common.RenderError(w, r, ErrorInvalidRequest(authService.ErrEmailAlreadyExists))
-	case errors.Is(authErr.Err, authService.ErrUsernameAlreadyExists):
-		common.RenderError(w, r, ErrorInvalidRequest(authService.ErrUsernameAlreadyExists))
-	case errors.Is(authErr.Err, authService.ErrPasswordTooWeak):
-		common.RenderError(w, r, ErrorInvalidRequest(authService.ErrPasswordTooWeak))
-	default:
-		common.RenderError(w, r, ErrorInternalServer(err))
-	}
-}
-
-// buildAccountResponse constructs an AccountResponse from an Account model
-func buildAccountResponse(account *authModel.Account) *AccountResponse {
-	resp := &AccountResponse{
-		ID:     account.ID,
-		Email:  account.Email,
-		Active: account.Active,
-	}
-
-	if account.Username != nil {
-		resp.Username = *account.Username
-	}
-
-	roleNames := make([]string, 0, len(account.Roles))
-	for _, role := range account.Roles {
-		roleNames = append(roleNames, role.Name)
-	}
-	resp.Roles = roleNames
-
-	return resp
-}
-
-// refreshToken handles token refresh
-func (rs *Resource) refreshToken(w http.ResponseWriter, r *http.Request) {
-	// Get refresh token from context
-	refreshToken := jwt.RefreshTokenFromCtx(r.Context())
-
-	// Get IP address and user agent for audit logging
-	ipAddress := getClientIP(r)
-	userAgent := r.Header.Get(headerUserAgent)
-
-	accessToken, newRefreshToken, err := rs.AuthService.RefreshTokenWithAudit(r.Context(), refreshToken, ipAddress, userAgent)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidToken):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrInvalidToken))
-			case errors.Is(authErr.Err, authService.ErrTokenExpired):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrTokenExpired))
-			case errors.Is(authErr.Err, authService.ErrTokenNotFound):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrTokenNotFound))
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrAccountNotFound))
-			case errors.Is(authErr.Err, authService.ErrAccountInactive):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrAccountInactive))
-			default:
-				common.RenderError(w, r, ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	// Special case for token refresh endpoint - frontend expects direct token response
-	render.JSON(w, r, TokenResponse{
-		AccessToken:  accessToken,
-		RefreshToken: newRefreshToken,
-	})
-}
-
-// logout handles user logout
-func (rs *Resource) logout(w http.ResponseWriter, r *http.Request) {
-	// Get refresh token from context
-	refreshToken := jwt.RefreshTokenFromCtx(r.Context())
-
-	// Get IP address and user agent for audit logging
-	ipAddress := getClientIP(r)
-	userAgent := r.Header.Get(headerUserAgent)
-
-	err := rs.AuthService.LogoutWithAudit(r.Context(), refreshToken, ipAddress, userAgent)
-	if err != nil {
-		// Even if there's an error, we want to consider the logout successful from the client's perspective
-		// Log the error on the server side for debugging
-		log.Printf("Logout audit logging failed (client logout still successful): ip=%s, error=%v", ipAddress, err)
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// ChangePasswordRequest represents the change password request payload
-type ChangePasswordRequest struct {
-	CurrentPassword string `json:"current_password"`
-	NewPassword     string `json:"new_password"`
-	ConfirmPassword string `json:"confirm_password"`
-}
-
-// Bind validates the change password request
-func (req *ChangePasswordRequest) Bind(_ *http.Request) error {
-	return validation.ValidateStruct(req,
-		validation.Field(&req.CurrentPassword, validation.Required),
-		validation.Field(&req.NewPassword, validation.Required, validation.Length(8, 0)),
-		validation.Field(&req.ConfirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if req.NewPassword != req.ConfirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// changePassword handles password change
-func (rs *Resource) changePassword(w http.ResponseWriter, r *http.Request) {
-	req := &ChangePasswordRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	// Get user ID from JWT claims
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	err := rs.AuthService.ChangePassword(r.Context(), claims.ID, req.CurrentPassword, req.NewPassword)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidCredentials):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrInvalidCredentials))
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, ErrorUnauthorized(authService.ErrAccountNotFound))
-			case errors.Is(authErr.Err, authService.ErrPasswordTooWeak):
-				common.RenderError(w, r, ErrorInvalidRequest(authService.ErrPasswordTooWeak))
-			default:
-				common.RenderError(w, r, ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// getAccount returns the current user's account details
-func (rs *Resource) getAccount(w http.ResponseWriter, r *http.Request) {
-	// Get user ID and permissions from JWT claims
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	account, err := rs.AuthService.GetAccountByID(r.Context(), claims.ID)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			if errors.Is(authErr.Err, authService.ErrAccountNotFound) {
-				common.RenderError(w, r, ErrorNotFound(authService.ErrAccountNotFound))
-				return
-			}
-		}
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	// Convert account to response
-	resp := &AccountResponse{
-		ID:     account.ID,
-		Email:  account.Email,
-		Active: account.Active,
-	}
-
-	if account.Username != nil {
-		resp.Username = *account.Username
-	}
-
-	roleNames := make([]string, 0, len(account.Roles))
-	for _, role := range account.Roles {
-		roleNames = append(roleNames, role.Name)
-	}
-	resp.Roles = roleNames
-
-	// Include permissions from JWT claims
-	resp.Permissions = claims.Permissions
-
-	common.Respond(w, r, http.StatusOK, resp, "Account information retrieved successfully")
-}
-
-// Role Management Request/Response Types
-
-// CreateRoleRequest represents the create role request payload
-type CreateRoleRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
-
-// Bind validates the create role request
-func (req *CreateRoleRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-	)
-}
-
-// UpdateRoleRequest represents the update role request payload
-type UpdateRoleRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
-
-// Bind validates the update role request
-func (req *UpdateRoleRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-	)
-}
-
-// RoleResponse represents a role response
-type RoleResponse struct {
-	ID          int64    `json:"id"`
-	Name        string   `json:"name"`
-	Description string   `json:"description"`
-	CreatedAt   string   `json:"created_at"`
-	UpdatedAt   string   `json:"updated_at"`
-	Permissions []string `json:"permissions,omitempty"`
-}
-
-// Permission Management Request/Response Types
-
-// CreatePermissionRequest represents the create permission request payload
-type CreatePermissionRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Resource    string `json:"resource"`
-	Action      string `json:"action"`
-}
-
-// Bind validates the create permission request
-func (req *CreatePermissionRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-	req.Resource = strings.TrimSpace(strings.ToLower(req.Resource))
-	req.Action = strings.TrimSpace(strings.ToLower(req.Action))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-		validation.Field(&req.Resource, validation.Required, validation.Length(1, 50)),
-		validation.Field(&req.Action, validation.Required, validation.Length(1, 50)),
-	)
-}
-
-// UpdatePermissionRequest represents the update permission request payload
-type UpdatePermissionRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Resource    string `json:"resource"`
-	Action      string `json:"action"`
-}
-
-// Bind validates the update permission request
-func (req *UpdatePermissionRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-	req.Resource = strings.TrimSpace(strings.ToLower(req.Resource))
-	req.Action = strings.TrimSpace(strings.ToLower(req.Action))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-		validation.Field(&req.Resource, validation.Required, validation.Length(1, 50)),
-		validation.Field(&req.Action, validation.Required, validation.Length(1, 50)),
-	)
-}
-
-// PermissionResponse represents a permission response
-type PermissionResponse struct {
-	ID          int64  `json:"id"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Resource    string `json:"resource"`
-	Action      string `json:"action"`
-	CreatedAt   string `json:"created_at"`
-	UpdatedAt   string `json:"updated_at"`
-}
-
-// UpdateAccountRequest represents the update account request payload
-type UpdateAccountRequest struct {
-	Email    string `json:"email"`
-	Username string `json:"username,omitempty"`
-}
-
-// Bind validates the update account request
-func (req *UpdateAccountRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-	req.Username = strings.TrimSpace(req.Username)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-		validation.Field(&req.Username, validation.Length(3, 30)),
-	)
-}
-
-// PasswordResetRequest represents the password reset request payload
-type PasswordResetRequest struct {
-	Email string `json:"email"`
-}
-
-// Bind validates the password reset request
-func (req *PasswordResetRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-	)
-}
-
-// PasswordResetConfirmRequest represents the password reset confirm request payload
-type PasswordResetConfirmRequest struct {
-	Token           string `json:"token"`
-	NewPassword     string `json:"new_password"`
-	ConfirmPassword string `json:"confirm_password"`
-}
-
-// Bind validates the password reset confirm request
-func (req *PasswordResetConfirmRequest) Bind(_ *http.Request) error {
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Token, validation.Required),
-		validation.Field(&req.NewPassword, validation.Required, validation.Length(8, 0)),
-		validation.Field(&req.ConfirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if req.NewPassword != req.ConfirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// CreateParentAccountRequest represents the create parent account request payload
-type CreateParentAccountRequest struct {
-	Email           string `json:"email"`
-	Username        string `json:"username"`
-	Password        string `json:"password"`
-	ConfirmPassword string `json:"confirm_password"`
-}
-
-// Bind validates the create parent account request
-func (req *CreateParentAccountRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-	req.Username = strings.TrimSpace(req.Username)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-		validation.Field(&req.Username, validation.Required, validation.Length(3, 30)),
-		validation.Field(&req.Password, validation.Required, validation.Length(8, 0)),
-		validation.Field(&req.ConfirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if req.Password != req.ConfirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// ParentAccountResponse represents a parent account response
-type ParentAccountResponse struct {
-	ID        int64  `json:"id"`
-	Email     string `json:"email"`
-	Username  string `json:"username,omitempty"`
-	Active    bool   `json:"active"`
-	CreatedAt string `json:"created_at"`
-	UpdatedAt string `json:"updated_at"`
-}
-
-// Role Management Endpoints
-
-// createRole handles creating a new role
-func (rs *Resource) createRole(w http.ResponseWriter, r *http.Request) {
-	req := &CreateRoleRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	role, err := rs.AuthService.CreateRole(r.Context(), req.Name, req.Description)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	resp := &RoleResponse{
-		ID:          role.ID,
-		Name:        role.Name,
-		Description: role.Description,
-		CreatedAt:   role.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   role.UpdatedAt.Format(time.RFC3339),
-	}
-
-	common.Respond(w, r, http.StatusCreated, resp, "Role created successfully")
-}
-
-// getRoleByID handles getting a role by ID
-func (rs *Resource) getRoleByID(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	role, err := rs.AuthService.GetRoleByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("role not found")))
-		return
-	}
-
-	// Get permissions for the role
-	permissions, _ := rs.AuthService.GetRolePermissions(r.Context(), id)
-	permissionNames := make([]string, 0, len(permissions))
-	for _, perm := range permissions {
-		permissionNames = append(permissionNames, perm.GetFullName())
-	}
-
-	resp := &RoleResponse{
-		ID:          role.ID,
-		Name:        role.Name,
-		Description: role.Description,
-		CreatedAt:   role.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   role.UpdatedAt.Format(time.RFC3339),
-		Permissions: permissionNames,
-	}
-
-	common.Respond(w, r, http.StatusOK, resp, "Role retrieved successfully")
-}
-
-// updateRole handles updating a role
-func (rs *Resource) updateRole(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	req := &UpdateRoleRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	role, err := rs.AuthService.GetRoleByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("role not found")))
-		return
-	}
-
-	role.Name = req.Name
-	role.Description = req.Description
-
-	if err := rs.AuthService.UpdateRole(r.Context(), role); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// deleteRole handles deleting a role
-func (rs *Resource) deleteRole(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DeleteRole(r.Context(), id); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listRoles handles listing roles
-func (rs *Resource) listRoles(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if name := r.URL.Query().Get("name"); name != "" {
-		filters["name"] = name
-	}
-
-	roles, err := rs.AuthService.ListRoles(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*RoleResponse, 0, len(roles))
-	for _, role := range roles {
-		resp := &RoleResponse{
-			ID:          role.ID,
-			Name:        role.Name,
-			Description: role.Description,
-			CreatedAt:   role.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   role.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Roles retrieved successfully")
-}
-
-// assignRoleToAccount handles assigning a role to an account
-func (rs *Resource) assignRoleToAccount(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	roleID, ok := common.ParseIntIDWithError(w, r, "roleId", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.AssignRoleToAccount(r.Context(), accountID, roleID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// removeRoleFromAccount handles removing a role from an account
-func (rs *Resource) removeRoleFromAccount(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	roleID, ok := common.ParseIntIDWithError(w, r, "roleId", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.RemoveRoleFromAccount(r.Context(), accountID, roleID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// getAccountRoles handles getting roles for an account
-func (rs *Resource) getAccountRoles(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	roles, err := rs.AuthService.GetAccountRoles(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*RoleResponse, 0, len(roles))
-	for _, role := range roles {
-		resp := &RoleResponse{
-			ID:          role.ID,
-			Name:        role.Name,
-			Description: role.Description,
-			CreatedAt:   role.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   role.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account roles retrieved successfully")
-}
-
-// Permission Management Endpoints
-
-// createPermission handles creating a new permission
-func (rs *Resource) createPermission(w http.ResponseWriter, r *http.Request) {
-	req := &CreatePermissionRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	permission, err := rs.AuthService.CreatePermission(r.Context(), req.Name, req.Description, req.Resource, req.Action)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	resp := &PermissionResponse{
-		ID:          permission.ID,
-		Name:        permission.Name,
-		Description: permission.Description,
-		Resource:    permission.Resource,
-		Action:      permission.Action,
-		CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-	}
-
-	common.Respond(w, r, http.StatusCreated, resp, "Permission created successfully")
-}
-
-// getPermissionByID handles getting a permission by ID
-func (rs *Resource) getPermissionByID(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	permission, err := rs.AuthService.GetPermissionByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("permission not found")))
-		return
-	}
-
-	resp := &PermissionResponse{
-		ID:          permission.ID,
-		Name:        permission.Name,
-		Description: permission.Description,
-		Resource:    permission.Resource,
-		Action:      permission.Action,
-		CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-	}
-
-	common.Respond(w, r, http.StatusOK, resp, "Permission retrieved successfully")
-}
-
-// updatePermission handles updating a permission
-func (rs *Resource) updatePermission(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	req := &UpdatePermissionRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	permission, err := rs.AuthService.GetPermissionByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("permission not found")))
-		return
-	}
-
-	permission.Name = req.Name
-	permission.Description = req.Description
-	permission.Resource = req.Resource
-	permission.Action = req.Action
-
-	if err := rs.AuthService.UpdatePermission(r.Context(), permission); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// deletePermission handles deleting a permission
-func (rs *Resource) deletePermission(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DeletePermission(r.Context(), id); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listPermissions handles listing permissions
-func (rs *Resource) listPermissions(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if resource := r.URL.Query().Get("resource"); resource != "" {
-		filters["resource"] = resource
-	}
-
-	if action := r.URL.Query().Get("action"); action != "" {
-		filters["action"] = action
-	}
-
-	permissions, err := rs.AuthService.ListPermissions(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Permissions retrieved successfully")
-}
-
-// grantPermissionToAccount handles granting a permission to an account
-func (rs *Resource) grantPermissionToAccount(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissionID, ok := common.ParseIntIDWithError(w, r, "permissionId", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.GrantPermissionToAccount(r.Context(), accountID, permissionID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// denyPermissionToAccount handles denying a permission to an account
-func (rs *Resource) denyPermissionToAccount(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissionID, ok := common.ParseIntIDWithError(w, r, "permissionId", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DenyPermissionToAccount(r.Context(), accountID, permissionID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// removePermissionFromAccount handles removing a permission from an account
-func (rs *Resource) removePermissionFromAccount(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissionID, ok := common.ParseIntIDWithError(w, r, "permissionId", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.RemovePermissionFromAccount(r.Context(), accountID, permissionID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// getAccountPermissions handles getting permissions for an account
-func (rs *Resource) getAccountPermissions(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissions, err := rs.AuthService.GetAccountPermissions(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account permissions retrieved successfully")
-}
-
-// getAccountDirectPermissions handles getting only direct permissions for an account (not role-based)
-func (rs *Resource) getAccountDirectPermissions(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissions, err := rs.AuthService.GetAccountDirectPermissions(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account direct permissions retrieved successfully")
-}
-
-// assignPermissionToRole handles assigning a permission to a role
-func (rs *Resource) assignPermissionToRole(w http.ResponseWriter, r *http.Request) {
-	roleID, ok := common.ParseIntIDWithError(w, r, "roleId", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	permissionID, ok := common.ParseIntIDWithError(w, r, "permissionId", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.AssignPermissionToRole(r.Context(), roleID, permissionID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// removePermissionFromRole handles removing a permission from a role
-func (rs *Resource) removePermissionFromRole(w http.ResponseWriter, r *http.Request) {
-	roleID, ok := common.ParseIntIDWithError(w, r, "roleId", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	permissionID, ok := common.ParseIntIDWithError(w, r, "permissionId", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.RemovePermissionFromRole(r.Context(), roleID, permissionID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// getRolePermissions handles getting permissions for a role
-func (rs *Resource) getRolePermissions(w http.ResponseWriter, r *http.Request) {
-	roleID, ok := common.ParseIntIDWithError(w, r, "roleId", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	permissions, err := rs.AuthService.GetRolePermissions(r.Context(), roleID)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Role permissions retrieved successfully")
-}
-
-// Account Management Extension Endpoints
-
-// activateAccount handles activating an account
-func (rs *Resource) activateAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.ActivateAccount(r.Context(), id); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// deactivateAccount handles deactivating an account
-func (rs *Resource) deactivateAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DeactivateAccount(r.Context(), id); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// updateAccount handles updating an account
-func (rs *Resource) updateAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	req := &UpdateAccountRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	account, err := rs.AuthService.GetAccountByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("account not found")))
-		return
-	}
-
-	account.Email = req.Email
-	if req.Username != "" {
-		username := req.Username
-		account.Username = &username
-	}
-
-	if err := rs.AuthService.UpdateAccount(r.Context(), account); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listAccounts handles listing accounts
-func (rs *Resource) listAccounts(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if email := r.URL.Query().Get("email"); email != "" {
-		filters["email"] = email
-	}
-
-	if active := r.URL.Query().Get("active"); active != "" {
-		switch active {
-		case "true":
-			filters["active"] = true
-		case "false":
-			filters["active"] = false
-		}
-	}
-
-	accounts, err := rs.AuthService.ListAccounts(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*AccountResponse, 0, len(accounts))
-	for _, account := range accounts {
-		resp := &AccountResponse{
-			ID:     account.ID,
-			Email:  account.Email,
-			Active: account.Active,
-		}
-
-		if account.Username != nil {
-			resp.Username = *account.Username
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Accounts retrieved successfully")
-}
-
-// getAccountsByRole handles getting accounts by role
-func (rs *Resource) getAccountsByRole(w http.ResponseWriter, r *http.Request) {
-	roleName := chi.URLParam(r, "roleName")
-
-	accounts, err := rs.AuthService.GetAccountsByRole(r.Context(), roleName)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*AccountResponse, 0, len(accounts))
-	for _, account := range accounts {
-		resp := &AccountResponse{
-			ID:     account.ID,
-			Email:  account.Email,
-			Active: account.Active,
-		}
-
-		if account.Username != nil {
-			resp.Username = *account.Username
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Accounts retrieved successfully")
-}
-
-// Password Reset Endpoints
-
-// initiatePasswordReset handles initiating a password reset
-func (rs *Resource) initiatePasswordReset(w http.ResponseWriter, r *http.Request) {
-	req := &PasswordResetRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	// Always return success to avoid revealing whether email exists, but handle rate limiting
-	_, err := rs.AuthService.InitiatePasswordReset(r.Context(), req.Email)
-	if err != nil {
-		var rateErr *authService.RateLimitError
-		if errors.As(err, &rateErr) {
-			// Prefer Retry-After seconds, fallback to RFC1123 format
-			retryAfterSeconds := rateErr.RetryAfterSeconds(time.Now())
-			if retryAfterSeconds > 0 {
-				w.Header().Set("Retry-After", strconv.Itoa(retryAfterSeconds))
-			} else if !rateErr.RetryAt.IsZero() {
-				w.Header().Set("Retry-After", rateErr.RetryAt.UTC().Format(http.TimeFormat))
-			}
-
-			common.RenderError(w, r, common.ErrorTooManyRequests(authService.ErrRateLimitExceeded))
-			return
-		}
-
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	log.Printf("Password reset initiated for email=%s", req.Email)
-
-	common.Respond(w, r, http.StatusOK, nil, "If the email exists, a password reset link has been sent")
-}
-
-// resetPassword handles resetting password with token
-func (rs *Resource) resetPassword(w http.ResponseWriter, r *http.Request) {
-	req := &PasswordResetConfirmRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	if err := rs.AuthService.ResetPassword(r.Context(), req.Token, req.NewPassword); err != nil {
-		log.Printf("Password reset failed reason=%v", err)
-
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidToken),
-				errors.Is(authErr.Err, sql.ErrNoRows):
-				// Both cases indicate the token is invalid or not found
-				common.RenderError(w, r, ErrorInvalidRequest(errors.New("invalid or expired reset token")))
-				return
-			case errors.Is(authErr.Err, authService.ErrPasswordTooWeak):
-				common.RenderError(w, r, ErrorInvalidRequest(authService.ErrPasswordTooWeak))
-				return
-			}
-		}
-
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	log.Printf("Password reset completed successfully")
-
-	common.Respond(w, r, http.StatusOK, nil, "Password reset successfully")
-}
-
-// Token Management Endpoints
-
-// cleanupExpiredTokens handles cleanup of expired tokens
-func (rs *Resource) cleanupExpiredTokens(w http.ResponseWriter, r *http.Request) {
-	count, err := rs.AuthService.CleanupExpiredTokens(r.Context())
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	response := map[string]int{"cleaned_tokens": count}
-	common.Respond(w, r, http.StatusOK, response, "Expired tokens cleaned up successfully")
-}
-
-// revokeAllTokens handles revoking all tokens for an account
-func (rs *Resource) revokeAllTokens(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.RevokeAllTokens(r.Context(), accountID); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// getActiveTokens handles getting active tokens for an account
-func (rs *Resource) getActiveTokens(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	tokens, err := rs.AuthService.GetActiveTokens(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	type TokenResponse struct {
-		ID         int64  `json:"id"`
-		Token      string `json:"token"`
-		Expiry     string `json:"expiry"`
-		Mobile     bool   `json:"mobile"`
-		Identifier string `json:"identifier,omitempty"`
-		CreatedAt  string `json:"created_at"`
-	}
-
-	responses := make([]*TokenResponse, 0, len(tokens))
-	for _, token := range tokens {
-		resp := &TokenResponse{
-			ID:        token.ID,
-			Token:     token.Token,
-			Expiry:    token.Expiry.Format(time.RFC3339),
-			Mobile:    token.Mobile,
-			CreatedAt: token.CreatedAt.Format(time.RFC3339),
-		}
-
-		if token.Identifier != nil {
-			resp.Identifier = *token.Identifier
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Active tokens retrieved successfully")
-}
-
-// Parent Account Management Endpoints
-
-// createParentAccount handles creating a parent account
-func (rs *Resource) createParentAccount(w http.ResponseWriter, r *http.Request) {
-	req := &CreateParentAccountRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	parentAccount, err := rs.AuthService.CreateParentAccount(r.Context(), req.Email, req.Username, req.Password)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrEmailAlreadyExists):
-				common.RenderError(w, r, ErrorInvalidRequest(authService.ErrEmailAlreadyExists))
-			case errors.Is(authErr.Err, authService.ErrUsernameAlreadyExists):
-				common.RenderError(w, r, ErrorInvalidRequest(authService.ErrUsernameAlreadyExists))
-			default:
-				common.RenderError(w, r, ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	resp := &ParentAccountResponse{
-		ID:        parentAccount.ID,
-		Email:     parentAccount.Email,
-		Active:    parentAccount.Active,
-		CreatedAt: parentAccount.CreatedAt.Format(time.RFC3339),
-		UpdatedAt: parentAccount.UpdatedAt.Format(time.RFC3339),
-	}
-
-	if parentAccount.Username != nil {
-		resp.Username = *parentAccount.Username
-	}
-
-	common.Respond(w, r, http.StatusCreated, resp, "Parent account created successfully")
-}
-
-// getParentAccountByID handles getting a parent account by ID
-func (rs *Resource) getParentAccountByID(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidParentAccountID)
-	if !ok {
-		return
-	}
-
-	parentAccount, err := rs.AuthService.GetParentAccountByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("parent account not found")))
-		return
-	}
-
-	resp := &ParentAccountResponse{
-		ID:        parentAccount.ID,
-		Email:     parentAccount.Email,
-		Active:    parentAccount.Active,
-		CreatedAt: parentAccount.CreatedAt.Format(time.RFC3339),
-		UpdatedAt: parentAccount.UpdatedAt.Format(time.RFC3339),
-	}
-
-	if parentAccount.Username != nil {
-		resp.Username = *parentAccount.Username
-	}
-
-	common.Respond(w, r, http.StatusOK, resp, "Parent account retrieved successfully")
-}
-
-// updateParentAccount handles updating a parent account
-func (rs *Resource) updateParentAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidParentAccountID)
-	if !ok {
-		return
-	}
-
-	req := &UpdateAccountRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, ErrorInvalidRequest(err))
-		return
-	}
-
-	parentAccount, err := rs.AuthService.GetParentAccountByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, ErrorNotFound(errors.New("parent account not found")))
-		return
-	}
-
-	parentAccount.Email = req.Email
-	if req.Username != "" {
-		username := req.Username
-		parentAccount.Username = &username
-	}
-
-	if err := rs.AuthService.UpdateParentAccount(r.Context(), parentAccount); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// activateParentAccount handles activating a parent account
-func (rs *Resource) activateParentAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidParentAccountID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.ActivateParentAccount(r.Context(), id); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// deactivateParentAccount handles deactivating a parent account
-func (rs *Resource) deactivateParentAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidParentAccountID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DeactivateParentAccount(r.Context(), id); err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listParentAccounts handles listing parent accounts
-func (rs *Resource) listParentAccounts(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if email := r.URL.Query().Get("email"); email != "" {
-		filters["email"] = email
-	}
-
-	if active := r.URL.Query().Get("active"); active != "" {
-		switch active {
-		case "true":
-			filters["active"] = true
-		case "false":
-			filters["active"] = false
-		}
-	}
-
-	parentAccounts, err := rs.AuthService.ListParentAccounts(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*ParentAccountResponse, 0, len(parentAccounts))
-	for _, account := range parentAccounts {
-		resp := &ParentAccountResponse{
-			ID:        account.ID,
-			Email:     account.Email,
-			Active:    account.Active,
-			CreatedAt: account.CreatedAt.Format(time.RFC3339),
-			UpdatedAt: account.UpdatedAt.Format(time.RFC3339),
-		}
-
-		if account.Username != nil {
-			resp.Username = *account.Username
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Parent accounts retrieved successfully")
-}
-
 // getClientIP extracts the real client IP address from the request
 func getClientIP(r *http.Request) string {
 	// Check X-Real-IP header first (set by reverse proxy)
@@ -1751,211 +204,17 @@ func getClientIP(r *http.Request) string {
 	return ip
 }
 
-// ============================================================================
-// EXPORTED HANDLER METHODS (for testing)
-// ============================================================================
-
-// GetAccountHandler returns the getAccount handler for testing
-func (rs *Resource) GetAccountHandler() http.HandlerFunc {
-	return rs.getAccount
+// trimAndLowerEmail trims whitespace and converts email to lowercase
+func trimAndLowerEmail(email string) string {
+	return strings.TrimSpace(strings.ToLower(email))
 }
 
-// ChangePasswordHandler returns the changePassword handler for testing
-func (rs *Resource) ChangePasswordHandler() http.HandlerFunc {
-	return rs.changePassword
-}
-
-// ListRolesHandler returns the listRoles handler for testing
-func (rs *Resource) ListRolesHandler() http.HandlerFunc {
-	return rs.listRoles
-}
-
-// CreateRoleHandler returns the createRole handler for testing
-func (rs *Resource) CreateRoleHandler() http.HandlerFunc {
-	return rs.createRole
-}
-
-// GetRoleByIDHandler returns the getRoleByID handler for testing
-func (rs *Resource) GetRoleByIDHandler() http.HandlerFunc {
-	return rs.getRoleByID
-}
-
-// DeleteRoleHandler returns the deleteRole handler for testing
-func (rs *Resource) DeleteRoleHandler() http.HandlerFunc {
-	return rs.deleteRole
-}
-
-// ListPermissionsHandler returns the listPermissions handler for testing
-func (rs *Resource) ListPermissionsHandler() http.HandlerFunc {
-	return rs.listPermissions
-}
-
-// CreatePermissionHandler returns the createPermission handler for testing
-func (rs *Resource) CreatePermissionHandler() http.HandlerFunc {
-	return rs.createPermission
-}
-
-// GetPermissionByIDHandler returns the getPermissionByID handler for testing
-func (rs *Resource) GetPermissionByIDHandler() http.HandlerFunc {
-	return rs.getPermissionByID
-}
-
-// ListAccountsHandler returns the listAccounts handler for testing
-func (rs *Resource) ListAccountsHandler() http.HandlerFunc {
-	return rs.listAccounts
-}
-
-// UpdateRoleHandler returns the updateRole handler for testing
-func (rs *Resource) UpdateRoleHandler() http.HandlerFunc {
-	return rs.updateRole
-}
-
-// UpdatePermissionHandler returns the updatePermission handler for testing
-func (rs *Resource) UpdatePermissionHandler() http.HandlerFunc {
-	return rs.updatePermission
-}
-
-// DeletePermissionHandler returns the deletePermission handler for testing
-func (rs *Resource) DeletePermissionHandler() http.HandlerFunc {
-	return rs.deletePermission
-}
-
-// AssignRoleToAccountHandler returns the assignRoleToAccount handler for testing
-func (rs *Resource) AssignRoleToAccountHandler() http.HandlerFunc {
-	return rs.assignRoleToAccount
-}
-
-// RemoveRoleFromAccountHandler returns the removeRoleFromAccount handler for testing
-func (rs *Resource) RemoveRoleFromAccountHandler() http.HandlerFunc {
-	return rs.removeRoleFromAccount
-}
-
-// GetAccountRolesHandler returns the getAccountRoles handler for testing
-func (rs *Resource) GetAccountRolesHandler() http.HandlerFunc {
-	return rs.getAccountRoles
-}
-
-// AssignPermissionToRoleHandler returns the assignPermissionToRole handler for testing
-func (rs *Resource) AssignPermissionToRoleHandler() http.HandlerFunc {
-	return rs.assignPermissionToRole
-}
-
-// RemovePermissionFromRoleHandler returns the removePermissionFromRole handler for testing
-func (rs *Resource) RemovePermissionFromRoleHandler() http.HandlerFunc {
-	return rs.removePermissionFromRole
-}
-
-// GetRolePermissionsHandler returns the getRolePermissions handler for testing
-func (rs *Resource) GetRolePermissionsHandler() http.HandlerFunc {
-	return rs.getRolePermissions
-}
-
-// GrantPermissionToAccountHandler returns the grantPermissionToAccount handler for testing
-func (rs *Resource) GrantPermissionToAccountHandler() http.HandlerFunc {
-	return rs.grantPermissionToAccount
-}
-
-// DenyPermissionToAccountHandler returns the denyPermissionToAccount handler for testing
-func (rs *Resource) DenyPermissionToAccountHandler() http.HandlerFunc {
-	return rs.denyPermissionToAccount
-}
-
-// RemovePermissionFromAccountHandler returns the removePermissionFromAccount handler for testing
-func (rs *Resource) RemovePermissionFromAccountHandler() http.HandlerFunc {
-	return rs.removePermissionFromAccount
-}
-
-// GetAccountPermissionsHandler returns the getAccountPermissions handler for testing
-func (rs *Resource) GetAccountPermissionsHandler() http.HandlerFunc {
-	return rs.getAccountPermissions
-}
-
-// GetAccountDirectPermissionsHandler returns the getAccountDirectPermissions handler for testing
-func (rs *Resource) GetAccountDirectPermissionsHandler() http.HandlerFunc {
-	return rs.getAccountDirectPermissions
-}
-
-// ActivateAccountHandler returns the activateAccount handler for testing
-func (rs *Resource) ActivateAccountHandler() http.HandlerFunc {
-	return rs.activateAccount
-}
-
-// DeactivateAccountHandler returns the deactivateAccount handler for testing
-func (rs *Resource) DeactivateAccountHandler() http.HandlerFunc {
-	return rs.deactivateAccount
-}
-
-// UpdateAccountHandler returns the updateAccount handler for testing
-func (rs *Resource) UpdateAccountHandler() http.HandlerFunc {
-	return rs.updateAccount
-}
-
-// GetAccountsByRoleHandler returns the getAccountsByRole handler for testing
-func (rs *Resource) GetAccountsByRoleHandler() http.HandlerFunc {
-	return rs.getAccountsByRole
-}
-
-// GetActiveTokensHandler returns the getActiveTokens handler for testing
-func (rs *Resource) GetActiveTokensHandler() http.HandlerFunc {
-	return rs.getActiveTokens
-}
-
-// RevokeAllTokensHandler returns the revokeAllTokens handler for testing
-func (rs *Resource) RevokeAllTokensHandler() http.HandlerFunc {
-	return rs.revokeAllTokens
-}
-
-// CleanupExpiredTokensHandler returns the cleanupExpiredTokens handler for testing
-func (rs *Resource) CleanupExpiredTokensHandler() http.HandlerFunc {
-	return rs.cleanupExpiredTokens
-}
-
-// CreateInvitationHandler returns the createInvitation handler for testing
-func (rs *Resource) CreateInvitationHandler() http.HandlerFunc {
-	return rs.createInvitation
-}
-
-// ListPendingInvitationsHandler returns the listPendingInvitations handler for testing
-func (rs *Resource) ListPendingInvitationsHandler() http.HandlerFunc {
-	return rs.listPendingInvitations
-}
-
-// ResendInvitationHandler returns the resendInvitation handler for testing
-func (rs *Resource) ResendInvitationHandler() http.HandlerFunc {
-	return rs.resendInvitation
-}
-
-// RevokeInvitationHandler returns the revokeInvitation handler for testing
-func (rs *Resource) RevokeInvitationHandler() http.HandlerFunc {
-	return rs.revokeInvitation
-}
-
-// CreateParentAccountHandler returns the createParentAccount handler for testing
-func (rs *Resource) CreateParentAccountHandler() http.HandlerFunc {
-	return rs.createParentAccount
-}
-
-// GetParentAccountByIDHandler returns the getParentAccountByID handler for testing
-func (rs *Resource) GetParentAccountByIDHandler() http.HandlerFunc {
-	return rs.getParentAccountByID
-}
-
-// UpdateParentAccountHandler returns the updateParentAccount handler for testing
-func (rs *Resource) UpdateParentAccountHandler() http.HandlerFunc {
-	return rs.updateParentAccount
-}
-
-// ActivateParentAccountHandler returns the activateParentAccount handler for testing
-func (rs *Resource) ActivateParentAccountHandler() http.HandlerFunc {
-	return rs.activateParentAccount
-}
-
-// DeactivateParentAccountHandler returns the deactivateParentAccount handler for testing
-func (rs *Resource) DeactivateParentAccountHandler() http.HandlerFunc {
-	return rs.deactivateParentAccount
-}
-
-// ListParentAccountsHandler returns the listParentAccounts handler for testing
-func (rs *Resource) ListParentAccountsHandler() http.HandlerFunc {
-	return rs.listParentAccounts
+// validateEmail validates an email address format
+func validateEmail(value interface{}) error {
+	email, _ := value.(string)
+	emailRegex := regexp.MustCompile(`^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$`)
+	if !emailRegex.MatchString(email) {
+		return ErrInvalidRequest
+	}
+	return nil
 }
